+++ blockapps-rest/ReleaseNotes.md
@@ -3,6 +3,7 @@
 ### Version 7.5.0
 
 -   Add support for resource-owner-password-credential grant flow
+-   Support overriding config client id and client secret for getAccessTokenByClientSecret
 
 ### Version 7.4.4